+++ STM32CubeProject/Src/main.c
@@ -38,12 +38,313 @@
 
 /* USER CODE BEGIN Includes */
 
+#include "lcd.h"
+
 /* USER CODE END Includes */
 
 /* Private variables ---------------------------------------------------------*/
 
 /* USER CODE BEGIN PV */
 /* Private variables ---------------------------------------------------------*/
+
+const unsigned char FONT6x8[97][8] = {
+0x06,0x08,0x08,0x00,0x00,0x00,0x00,0x00,            // Columns, rows, num_bytes_per_char
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,            // space  0x20
+0x20,0x20,0x20,0x20,0x20,0x00,0x20,0x00,            // !
+0x50,0x50,0x50,0x00,0x00,0x00,0x00,0x00,            // "
+0x50,0x50,0xF8,0x50,0xF8,0x50,0x50,0x00,            // #
+0x20,0x78,0xA0,0x70,0x28,0xF0,0x20,0x00,            // $
+0xC0,0xC8,0x10,0x20,0x40,0x98,0x18,0x00,            // %
+0x40,0xA0,0xA0,0x40,0xA8,0x90,0x68,0x00,            // &
+0x30,0x30,0x20,0x40,0x00,0x00,0x00,0x00,            // '
+0x10,0x20,0x40,0x40,0x40,0x20,0x10,0x00,            // (
+0x40,0x20,0x10,0x10,0x10,0x20,0x40,0x00,            // )
+0x00,0x20,0xA8,0x70,0x70,0xA8,0x20,0x00,            // *
+0x00,0x20,0x20,0xF8,0x20,0x20,0x00,0x00,            // +
+0x00,0x00,0x00,0x00,0x30,0x30,0x20,0x40,            // ,
+0x00,0x00,0x00,0xF8,0x00,0x00,0x00,0x00,            // -
+0x00,0x00,0x00,0x00,0x00,0x30,0x30,0x00,            // .
+0x00,0x08,0x10,0x20,0x40,0x80,0x00,0x00,            // / (forward slash)
+0x70,0x88,0x88,0xA8,0x88,0x88,0x70,0x00,            // 0
+0x20,0x60,0x20,0x20,0x20,0x20,0x70,0x00,            // 1
+0x70,0x88,0x08,0x70,0x80,0x80,0xF8,0x00,            // 2
+0xF8,0x08,0x10,0x30,0x08,0x88,0x70,0x00,            // 3
+0x10,0x30,0x50,0x90,0xF8,0x10,0x10,0x00,            // 4
+0xF8,0x80,0xF0,0x08,0x08,0x88,0x70,0x00,            // 5
+0x38,0x40,0x80,0xF0,0x88,0x88,0x70,0x00,            // 6
+0xF8,0x08,0x08,0x10,0x20,0x40,0x80,0x00,            // 7
+0x70,0x88,0x88,0x70,0x88,0x88,0x70,0x00,            // 8
+0x70,0x88,0x88,0x78,0x08,0x10,0xE0,0x00,            // 9
+0x00,0x00,0x20,0x00,0x20,0x00,0x00,0x00,            // :
+0x00,0x00,0x20,0x00,0x20,0x20,0x40,0x00,            // ;
+0x08,0x10,0x20,0x40,0x20,0x10,0x08,0x00,            // <
+0x00,0x00,0xF8,0x00,0xF8,0x00,0x00,0x00,            // =
+0x40,0x20,0x10,0x08,0x10,0x20,0x40,0x00,            // >
+0x70,0x88,0x08,0x30,0x20,0x00,0x20,0x00,            // ?
+0x70,0x88,0xA8,0xB8,0xB0,0x80,0x78,0x00,            // @
+0x20,0x50,0x88,0x88,0xF8,0x88,0x88,0x00,            // A
+0xF0,0x88,0x88,0xF0,0x88,0x88,0xF0,0x00,            // B
+0x70,0x88,0x80,0x80,0x80,0x88,0x70,0x00,            // C
+0xF0,0x88,0x88,0x88,0x88,0x88,0xF0,0x00,            // D
+0xF8,0x80,0x80,0xF0,0x80,0x80,0xF8,0x00,            // E
+0xF8,0x80,0x80,0xF0,0x80,0x80,0x80,0x00,            // F
+0x78,0x88,0x80,0x80,0x98,0x88,0x78,0x00,            // G
+0x88,0x88,0x88,0xF8,0x88,0x88,0x88,0x00,            // H
+0x70,0x20,0x20,0x20,0x20,0x20,0x70,0x00,            // I
+0x38,0x10,0x10,0x10,0x10,0x90,0x60,0x00,            // J
+0x88,0x90,0xA0,0xC0,0xA0,0x90,0x88,0x00,            // K
+0x80,0x80,0x80,0x80,0x80,0x80,0xF8,0x00,            // L
+0x88,0xD8,0xA8,0xA8,0xA8,0x88,0x88,0x00,            // M
+0x88,0x88,0xC8,0xA8,0x98,0x88,0x88,0x00,            // N
+0x70,0x88,0x88,0x88,0x88,0x88,0x70,0x00,            // O
+0xF0,0x88,0x88,0xF0,0x80,0x80,0x80,0x00,            // P
+0x70,0x88,0x88,0x88,0xA8,0x90,0x68,0x00,            // Q
+0xF0,0x88,0x88,0xF0,0xA0,0x90,0x88,0x00,            // R
+0x70,0x88,0x80,0x70,0x08,0x88,0x70,0x00,            // S
+0xF8,0xA8,0x20,0x20,0x20,0x20,0x20,0x00,            // T
+0x88,0x88,0x88,0x88,0x88,0x88,0x70,0x00,            // U
+0x88,0x88,0x88,0x88,0x88,0x50,0x20,0x00,            // V
+0x88,0x88,0x88,0xA8,0xA8,0xA8,0x50,0x00,            // W
+0x88,0x88,0x50,0x20,0x50,0x88,0x88,0x00,            // X
+0x88,0x88,0x50,0x20,0x20,0x20,0x20,0x00,            // Y
+0xF8,0x08,0x10,0x70,0x40,0x80,0xF8,0x00,            // Z
+0x78,0x40,0x40,0x40,0x40,0x40,0x78,0x00,            // [
+0x00,0x80,0x40,0x20,0x10,0x08,0x00,0x00,            // (back slash)
+0x78,0x08,0x08,0x08,0x08,0x08,0x78,0x00,            // ]
+0x20,0x50,0x88,0x00,0x00,0x00,0x00,0x00,            // ^
+0x00,0x00,0x00,0x00,0x00,0x00,0xF8,0x00,            // _
+0x60,0x60,0x20,0x10,0x00,0x00,0x00,0x00,            // `
+0x00,0x00,0x60,0x10,0x70,0x90,0x78,0x00,            // a
+0x80,0x80,0xB0,0xC8,0x88,0xC8,0xB0,0x00,            // b
+0x00,0x00,0x70,0x88,0x80,0x88,0x70,0x00,            // c
+0x08,0x08,0x68,0x98,0x88,0x98,0x68,0x00,            // d
+0x00,0x00,0x70,0x88,0xF8,0x80,0x70,0x00,            // e
+0x10,0x28,0x20,0x70,0x20,0x20,0x20,0x00,            // f
+0x00,0x00,0x70,0x98,0x98,0x68,0x08,0x70,            // g
+0x80,0x80,0xB0,0xC8,0x88,0x88,0x88,0x00,            // h
+0x20,0x00,0x60,0x20,0x20,0x20,0x70,0x00,            // i
+0x10,0x00,0x10,0x10,0x10,0x90,0x60,0x00,            // j
+0x80,0x80,0x90,0xA0,0xC0,0xA0,0x90,0x00,            // k
+0x60,0x20,0x20,0x20,0x20,0x20,0x70,0x00,            // l
+0x00,0x00,0xD0,0xA8,0xA8,0xA8,0xA8,0x00,            // m
+0x00,0x00,0xB0,0xC8,0x88,0x88,0x88,0x00,            // n
+0x00,0x00,0x70,0x88,0x88,0x88,0x70,0x00,            // o
+0x00,0x00,0xB0,0xC8,0xC8,0xB0,0x80,0x80,            // p
+0x00,0x00,0x68,0x98,0x98,0x68,0x08,0x08,            // q
+0x00,0x00,0xB0,0xC8,0x80,0x80,0x80,0x00,            // r
+0x00,0x00,0x78,0x80,0x70,0x08,0xF0,0x00,            // s
+0x20,0x20,0xF8,0x20,0x20,0x28,0x10,0x00,            // t
+0x00,0x00,0x88,0x88,0x88,0x98,0x68,0x00,            // u
+0x00,0x00,0x88,0x88,0x88,0x50,0x20,0x00,            // v
+0x00,0x00,0x88,0x88,0xA8,0xA8,0x50,0x00,            // w
+0x00,0x00,0x88,0x50,0x20,0x50,0x88,0x00,            // x
+0x00,0x00,0x88,0x88,0x78,0x08,0x88,0x70,            // y
+0x00,0x00,0xF8,0x10,0x20,0x40,0xF8,0x00,            // z
+0x10,0x20,0x20,0x40,0x20,0x20,0x10,0x00,            // {
+0x20,0x20,0x20,0x00,0x20,0x20,0x20,0x00,            // |
+0x40,0x20,0x20,0x10,0x20,0x20,0x40,0x00,            // }
+0x40,0xA8,0x10,0x00,0x00,0x00,0x00,0x00,            // ~
+0x70,0xD8,0xD8,0x70,0x00,0x00,0x00,0x00};           // DEL
+
+const unsigned char FONT8x8[97][8] = {
+0x08,0x08,0x08,0x00,0x00,0x00,0x00,0x00,  // columns, rows, num_bytes_per_char
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,  // space
+0x30,0x78,0x78,0x30,0x30,0x00,0x30,0x00,  // !
+0x6C,0x6C,0x6C,0x00,0x00,0x00,0x00,0x00,  // "
+0x6C,0x6C,0xFE,0x6C,0xFE,0x6C,0x6C,0x00,  // #
+0x18,0x3E,0x60,0x3C,0x06,0x7C,0x18,0x00,  // $
+0x00,0x63,0x66,0x0C,0x18,0x33,0x63,0x00,  // %
+0x1C,0x36,0x1C,0x3B,0x6E,0x66,0x3B,0x00,  // &
+0x30,0x30,0x60,0x00,0x00,0x00,0x00,0x00,  // '
+0x0C,0x18,0x30,0x30,0x30,0x18,0x0C,0x00,  // (
+0x30,0x18,0x0C,0x0C,0x0C,0x18,0x30,0x00,  // )
+0x00,0x66,0x3C,0xFF,0x3C,0x66,0x00,0x00,  // *
+0x00,0x30,0x30,0xFC,0x30,0x30,0x00,0x00,  // +
+0x00,0x00,0x00,0x00,0x00,0x18,0x18,0x30,  // ,
+0x00,0x00,0x00,0x7E,0x00,0x00,0x00,0x00,  // -
+0x00,0x00,0x00,0x00,0x00,0x18,0x18,0x00,  // .
+0x03,0x06,0x0C,0x18,0x30,0x60,0x40,0x00,  // / (forward slash)
+0x3E,0x63,0x63,0x6B,0x63,0x63,0x3E,0x00,  // 0
+0x18,0x38,0x58,0x18,0x18,0x18,0x7E,0x00,  // 1
+0x3C,0x66,0x06,0x1C,0x30,0x66,0x7E,0x00,  // 2
+0x3C,0x66,0x06,0x1C,0x06,0x66,0x3C,0x00,  // 3
+0x0E,0x1E,0x36,0x66,0x7F,0x06,0x0F,0x00,  // 4
+0x7E,0x60,0x7C,0x06,0x06,0x66,0x3C,0x00,  // 5
+0x1C,0x30,0x60,0x7C,0x66,0x66,0x3C,0x00,  // 6
+0x7E,0x66,0x06,0x0C,0x18,0x18,0x18,0x00,  // 7
+0x3C,0x66,0x66,0x3C,0x66,0x66,0x3C,0x00,  // 8
+0x3C,0x66,0x66,0x3E,0x06,0x0C,0x38,0x00,  // 9
+0x00,0x18,0x18,0x00,0x00,0x18,0x18,0x00,  // :
+0x00,0x18,0x18,0x00,0x00,0x18,0x18,0x30,  // ;
+0x0C,0x18,0x30,0x60,0x30,0x18,0x0C,0x00,  // <
+0x00,0x00,0x7E,0x00,0x00,0x7E,0x00,0x00,  // =
+0x30,0x18,0x0C,0x06,0x0C,0x18,0x30,0x00,  // >
+0x3C,0x66,0x06,0x0C,0x18,0x00,0x18,0x00,  // ?
+0x3E,0x63,0x6F,0x69,0x6F,0x60,0x3E,0x00,  // @
+0x18,0x3C,0x66,0x66,0x7E,0x66,0x66,0x00,  // A
+0x7E,0x33,0x33,0x3E,0x33,0x33,0x7E,0x00,  // B
+0x1E,0x33,0x60,0x60,0x60,0x33,0x1E,0x00,  // C
+0x7C,0x36,0x33,0x33,0x33,0x36,0x7C,0x00,  // D
+0x7F,0x31,0x34,0x3C,0x34,0x31,0x7F,0x00,  // E
+0x7F,0x31,0x34,0x3C,0x34,0x30,0x78,0x00,  // F
+0x1E,0x33,0x60,0x60,0x67,0x33,0x1F,0x00,  // G
+0x66,0x66,0x66,0x7E,0x66,0x66,0x66,0x00,  // H
+0x3C,0x18,0x18,0x18,0x18,0x18,0x3C,0x00,  // I
+0x0F,0x06,0x06,0x06,0x66,0x66,0x3C,0x00,  // J
+0x73,0x33,0x36,0x3C,0x36,0x33,0x73,0x00,  // K
+0x78,0x30,0x30,0x30,0x31,0x33,0x7F,0x00,  // L
+0x63,0x77,0x7F,0x7F,0x6B,0x63,0x63,0x00,  // M
+0x63,0x73,0x7B,0x6F,0x67,0x63,0x63,0x00,  // N
+0x3E,0x63,0x63,0x63,0x63,0x63,0x3E,0x00,  // O
+0x7E,0x33,0x33,0x3E,0x30,0x30,0x78,0x00,  // P
+0x3C,0x66,0x66,0x66,0x6E,0x3C,0x0E,0x00,  // Q
+0x7E,0x33,0x33,0x3E,0x36,0x33,0x73,0x00,  // R
+0x3C,0x66,0x30,0x18,0x0C,0x66,0x3C,0x00,  // S
+0x7E,0x5A,0x18,0x18,0x18,0x18,0x3C,0x00,  // T
+0x66,0x66,0x66,0x66,0x66,0x66,0x7E,0x00,  // U
+0x66,0x66,0x66,0x66,0x66,0x3C,0x18,0x00,  // V
+0x63,0x63,0x63,0x6B,0x7F,0x77,0x63,0x00,  // W
+0x63,0x63,0x36,0x1C,0x1C,0x36,0x63,0x00,  // X
+0x66,0x66,0x66,0x3C,0x18,0x18,0x3C,0x00,  // Y
+0x7F,0x63,0x46,0x0C,0x19,0x33,0x7F,0x00,  // Z
+0x3C,0x30,0x30,0x30,0x30,0x30,0x3C,0x00,  // [
+0x60,0x30,0x18,0x0C,0x06,0x03,0x01,0x00,  // (back slash)
+0x3C,0x0C,0x0C,0x0C,0x0C,0x0C,0x3C,0x00,  // ]
+0x08,0x1C,0x36,0x63,0x00,0x00,0x00,0x00,  // ^
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFF,  // _
+0x18,0x18,0x0C,0x00,0x00,0x00,0x00,0x00,   // `
+0x00,0x00,0x3C,0x06,0x3E,0x66,0x3B,0x00,  // a
+0x70,0x30,0x3E,0x33,0x33,0x33,0x6E,0x00,  // b
+0x00,0x00,0x3C,0x66,0x60,0x66,0x3C,0x00,  // c
+0x0E,0x06,0x3E,0x66,0x66,0x66,0x3B,0x00,  // d
+0x00,0x00,0x3C,0x66,0x7E,0x60,0x3C,0x00,  // e
+0x1C,0x36,0x30,0x78,0x30,0x30,0x78,0x00,  // f
+0x00,0x00,0x3B,0x66,0x66,0x3E,0x06,0x7C,  // g
+0x70,0x30,0x36,0x3B,0x33,0x33,0x73,0x00,  // h
+0x18,0x00,0x38,0x18,0x18,0x18,0x3C,0x00,  // i
+0x06,0x00,0x06,0x06,0x06,0x66,0x66,0x3C,  // j
+0x70,0x30,0x33,0x36,0x3C,0x36,0x73,0x00,  // k
+0x38,0x18,0x18,0x18,0x18,0x18,0x3C,0x00,  // l
+0x00,0x00,0x66,0x7F,0x7F,0x6B,0x63,0x00,  // m
+0x00,0x00,0x7C,0x66,0x66,0x66,0x66,0x00,  // n
+0x00,0x00,0x3C,0x66,0x66,0x66,0x3C,0x00,  // o
+0x00,0x00,0x6E,0x33,0x33,0x3E,0x30,0x78,  // p
+0x00,0x00,0x3B,0x66,0x66,0x3E,0x06,0x0F,  // q
+0x00,0x00,0x6E,0x3B,0x33,0x30,0x78,0x00,  // r
+0x00,0x00,0x3E,0x60,0x3C,0x06,0x7C,0x00,  // s
+0x08,0x18,0x3E,0x18,0x18,0x1A,0x0C,0x00,  // t
+0x00,0x00,0x66,0x66,0x66,0x66,0x3B,0x00,  // u
+0x00,0x00,0x66,0x66,0x66,0x3C,0x18,0x00,  // v
+0x00,0x00,0x63,0x6B,0x7F,0x7F,0x36,0x00,  // w
+0x00,0x00,0x63,0x36,0x1C,0x36,0x63,0x00,  // x
+0x00,0x00,0x66,0x66,0x66,0x3E,0x06,0x7C,  // y
+0x00,0x00,0x7E,0x4C,0x18,0x32,0x7E,0x00,  // z
+0x0E,0x18,0x18,0x70,0x18,0x18,0x0E,0x00,  // {
+0x0C,0x0C,0x0C,0x00,0x0C,0x0C,0x0C,0x00,  // |
+0x70,0x18,0x18,0x0E,0x18,0x18,0x70,0x00,  // }
+0x3B,0x6E,0x00,0x00,0x00,0x00,0x00,0x00,  // ~
+0x1C,0x36,0x36,0x1C,0x00,0x00,0x00,0x00   // DEL
+};
+
+const unsigned char FONT8x16[97][16] = {
+0x08,0x10,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x18,0x3C,0x3C,0x3C,0x18,0x18,0x18,0x00,0x18,0x18,0x00,0x00,0x00,0x00,
+0x00,0x63,0x63,0x63,0x22,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x36,0x36,0x7F,0x36,0x36,0x36,0x7F,0x36,0x36,0x00,0x00,0x00,0x00,
+0x0C,0x0C,0x3E,0x63,0x61,0x60,0x3E,0x03,0x03,0x43,0x63,0x3E,0x0C,0x0C,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x61,0x63,0x06,0x0C,0x18,0x33,0x63,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x1C,0x36,0x36,0x1C,0x3B,0x6E,0x66,0x66,0x3B,0x00,0x00,0x00,0x00,
+0x00,0x30,0x30,0x30,0x60,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x0C,0x18,0x18,0x30,0x30,0x30,0x30,0x18,0x18,0x0C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x18,0x0C,0x0C,0x06,0x06,0x06,0x06,0x0C,0x0C,0x18,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x42,0x66,0x3C,0xFF,0x3C,0x66,0x42,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x18,0x18,0x18,0xFF,0x18,0x18,0x18,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x18,0x18,0x18,0x30,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFF,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x18,0x18,0x00,0x00,0x00,0x00,
+0x00,0x00,0x01,0x03,0x07,0x0E,0x1C,0x38,0x70,0xE0,0xC0,0x80,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x63,0x63,0x6B,0x6B,0x63,0x63,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x0C,0x1C,0x3C,0x0C,0x0C,0x0C,0x0C,0x0C,0x0C,0x3F,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x03,0x06,0x0C,0x18,0x30,0x61,0x63,0x7F,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x03,0x03,0x1E,0x03,0x03,0x03,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x06,0x0E,0x1E,0x36,0x66,0x66,0x7F,0x06,0x06,0x0F,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7F,0x60,0x60,0x60,0x7E,0x03,0x03,0x63,0x73,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x1C,0x30,0x60,0x60,0x7E,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7F,0x63,0x03,0x06,0x06,0x0C,0x0C,0x18,0x18,0x18,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x63,0x63,0x3E,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x63,0x63,0x63,0x3F,0x03,0x03,0x06,0x3C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x18,0x18,0x00,0x00,0x00,0x18,0x18,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x18,0x18,0x00,0x00,0x00,0x18,0x18,0x18,0x30,0x00,0x00,
+0x00,0x00,0x00,0x06,0x0C,0x18,0x30,0x60,0x30,0x18,0x0C,0x06,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x7E,0x00,0x00,0x7E,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x60,0x30,0x18,0x0C,0x06,0x0C,0x18,0x30,0x60,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x63,0x06,0x0C,0x0C,0x0C,0x00,0x0C,0x0C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x63,0x6F,0x6B,0x6B,0x6E,0x60,0x60,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x08,0x1C,0x36,0x63,0x63,0x63,0x7F,0x63,0x63,0x63,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7E,0x33,0x33,0x33,0x3E,0x33,0x33,0x33,0x33,0x7E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x1E,0x33,0x61,0x60,0x60,0x60,0x60,0x61,0x33,0x1E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7C,0x36,0x33,0x33,0x33,0x33,0x33,0x33,0x36,0x7C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7F,0x33,0x31,0x34,0x3C,0x34,0x30,0x31,0x33,0x7F,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7F,0x33,0x31,0x34,0x3C,0x34,0x30,0x30,0x30,0x78,0x00,0x00,0x00,0x00,
+0x00,0x00,0x1E,0x33,0x61,0x60,0x60,0x6F,0x63,0x63,0x37,0x1D,0x00,0x00,0x00,0x00,
+0x00,0x00,0x63,0x63,0x63,0x63,0x7F,0x63,0x63,0x63,0x63,0x63,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3C,0x18,0x18,0x18,0x18,0x18,0x18,0x18,0x18,0x3C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x0F,0x06,0x06,0x06,0x06,0x06,0x06,0x66,0x66,0x3C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x73,0x33,0x36,0x36,0x3C,0x36,0x36,0x33,0x33,0x73,0x00,0x00,0x00,0x00,
+0x00,0x00,0x78,0x30,0x30,0x30,0x30,0x30,0x30,0x31,0x33,0x7F,0x00,0x00,0x00,0x00,
+0x00,0x00,0x63,0x77,0x7F,0x6B,0x63,0x63,0x63,0x63,0x63,0x63,0x00,0x00,0x00,0x00,
+0x00,0x00,0x63,0x63,0x73,0x7B,0x7F,0x6F,0x67,0x63,0x63,0x63,0x00,0x00,0x00,0x00,
+0x00,0x00,0x1C,0x36,0x63,0x63,0x63,0x63,0x63,0x63,0x36,0x1C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7E,0x33,0x33,0x33,0x3E,0x30,0x30,0x30,0x30,0x78,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x63,0x63,0x63,0x63,0x63,0x6B,0x6F,0x3E,0x06,0x07,0x00,0x00,
+0x00,0x00,0x7E,0x33,0x33,0x33,0x3E,0x36,0x36,0x33,0x33,0x73,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3E,0x63,0x63,0x30,0x1C,0x06,0x03,0x63,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0xFF,0xDB,0x99,0x18,0x18,0x18,0x18,0x18,0x18,0x3C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x63,0x63,0x63,0x63,0x63,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x63,0x63,0x63,0x63,0x63,0x63,0x63,0x36,0x1C,0x08,0x00,0x00,0x00,0x00,
+0x00,0x00,0x63,0x63,0x63,0x63,0x63,0x6B,0x6B,0x7F,0x36,0x36,0x00,0x00,0x00,0x00,
+0x00,0x00,0xC3,0xC3,0x66,0x3C,0x18,0x18,0x3C,0x66,0xC3,0xC3,0x00,0x00,0x00,0x00,
+0x00,0x00,0xC3,0xC3,0xC3,0x66,0x3C,0x18,0x18,0x18,0x18,0x3C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x7F,0x63,0x43,0x06,0x0C,0x18,0x30,0x61,0x63,0x7F,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3C,0x30,0x30,0x30,0x30,0x30,0x30,0x30,0x30,0x3C,0x00,0x00,0x00,0x00,
+0x00,0x00,0x80,0xC0,0xE0,0x70,0x38,0x1C,0x0E,0x07,0x03,0x01,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3C,0x0C,0x0C,0x0C,0x0C,0x0C,0x0C,0x0C,0x0C,0x3C,0x00,0x00,0x00,0x00,
+0x08,0x1C,0x36,0x63,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xFF,0x00,0x00,0x00,
+0x18,0x18,0x0C,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x3C,0x46,0x06,0x3E,0x66,0x66,0x3B,0x00,0x00,0x00,0x00,
+0x00,0x00,0x70,0x30,0x30,0x3C,0x36,0x33,0x33,0x33,0x33,0x6E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x3E,0x63,0x60,0x60,0x60,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x0E,0x06,0x06,0x1E,0x36,0x66,0x66,0x66,0x66,0x3B,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x3E,0x63,0x63,0x7E,0x60,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x1C,0x36,0x32,0x30,0x7C,0x30,0x30,0x30,0x30,0x78,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x3B,0x66,0x66,0x66,0x66,0x3E,0x06,0x66,0x3C,0x00,0x00,
+0x00,0x00,0x70,0x30,0x30,0x36,0x3B,0x33,0x33,0x33,0x33,0x73,0x00,0x00,0x00,0x00,
+0x00,0x00,0x0C,0x0C,0x00,0x1C,0x0C,0x0C,0x0C,0x0C,0x0C,0x1E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x06,0x06,0x00,0x0E,0x06,0x06,0x06,0x06,0x06,0x66,0x66,0x3C,0x00,0x00,
+0x00,0x00,0x70,0x30,0x30,0x33,0x33,0x36,0x3C,0x36,0x33,0x73,0x00,0x00,0x00,0x00,
+0x00,0x00,0x1C,0x0C,0x0C,0x0C,0x0C,0x0C,0x0C,0x0C,0x0C,0x1E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x6E,0x7F,0x6B,0x6B,0x6B,0x6B,0x6B,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x6E,0x33,0x33,0x33,0x33,0x33,0x33,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x3E,0x63,0x63,0x63,0x63,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x6E,0x33,0x33,0x33,0x33,0x3E,0x30,0x30,0x78,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x3B,0x66,0x66,0x66,0x66,0x3E,0x06,0x06,0x0F,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x6E,0x3B,0x33,0x30,0x30,0x30,0x78,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x3E,0x63,0x38,0x0E,0x03,0x63,0x3E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x08,0x18,0x18,0x7E,0x18,0x18,0x18,0x18,0x1B,0x0E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x66,0x66,0x66,0x66,0x66,0x66,0x3B,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x63,0x63,0x36,0x36,0x1C,0x1C,0x08,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x63,0x63,0x63,0x6B,0x6B,0x7F,0x36,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x63,0x36,0x1C,0x1C,0x1C,0x36,0x63,0x00,0x00,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x63,0x63,0x63,0x63,0x63,0x3F,0x03,0x06,0x3C,0x00,0x00,
+0x00,0x00,0x00,0x00,0x00,0x7F,0x66,0x0C,0x18,0x30,0x63,0x7F,0x00,0x00,0x00,0x00,
+0x00,0x00,0x0E,0x18,0x18,0x18,0x70,0x18,0x18,0x18,0x18,0x0E,0x00,0x00,0x00,0x00,
+0x00,0x00,0x18,0x18,0x18,0x18,0x18,0x00,0x18,0x18,0x18,0x18,0x18,0x00,0x00,0x00,
+0x00,0x00,0x70,0x18,0x18,0x18,0x0E,0x18,0x18,0x18,0x18,0x70,0x00,0x00,0x00,0x00,
+0x00,0x00,0x3B,0x6E,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
+0x00,0x70,0xD8,0xD8,0x70,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00
+};
 
 /* USER CODE END PV */
 
@@ -53,6 +354,290 @@ void Error_Handler(void);
 
 /* USER CODE BEGIN PFP */
 /* Private function prototypes -----------------------------------------------*/
+
+void write_spi_command(volatile unsigned int command) {
+  // clear bit 8 - indicates a "command"
+  command = (command & ~0x0100);
+  HAL_SPI_Transmit_IT(&hspi1, &command, SPI_DATASIZE_9BIT);
+}
+
+void write_spi_data(volatile unsigned int data) {
+  // set bit 8 - indicating a "command"
+  data = (data | 0x0100);
+  HAL_SPI_Transmit_IT(&hspi1, &command, SPI_DATASIZE_9BIT);
+}
+
+void backlight(unsigned char state) {
+  // The shield was build in a way the backlight cannot be mannualy controlled.
+  // So this does nothing
+}
+
+void init_lcd(void) {
+  // Hardware reset
+  LCD_RESET_LOW;
+  HAL_Delay(1);
+  LCD_RESET_HIGH;
+  HAL_Delay(1);
+  // Display control
+  write_spi_command(DISCTL);
+  write_spi_data(0x00); // P1: 0x00 = 2 divisions, switching period=8 (default) WriteSpiData(0x20); // P2: 0x20 = nlines/4 - 1 = 132/4 - 1 = 32) WriteSpiData(0x00); // P3: 0x00 = no inversely highlighted lines
+  // COM scan
+  write_spi_command(COMSCN);
+  write_spi_data(1);
+  // Internal oscilator ON
+  write_spi_command(OSCON);
+  // Sleep out
+  write_spi_command(SLPOUT);
+  // Power control
+  write_spi_command(PWRCTR);
+  write_spi_data(0x0f);   //
+  // Inverse display
+  write_spi_command(DISINV);
+}
+
+
+void lcd_write_130_130_bmp(void) {
+  // This function should get a picture defined in bmp.h and show it
+  // Because we won't be dealing with images, this does nothing
+
+}
+
+void lcd_clear_screen(void) {
+  long i;
+  // Row address set  (command 0x2B)
+  write_spi_command(PASET);
+  write_spi_data(0);
+  write_spi_data(131);
+  // loop counter
+  // Column address set  (command 0x2A)
+  write_spi_command(CASET);
+  write_spi_data(0);
+  write_spi_data(131);
+  // set the display memory to BLACK
+  write_spi_command(RAMWR);
+  for(i = 0; i < ((131 * 131) / 2); i++) {
+    write_spi_data((BLACK >> 4) & 0xFF);
+    write_spi_data(((BLACK & 0xF) << 4) | ((BLACK >> 8) & 0xF));
+    write_spi_data(BLACK & 0xFF);
+  }
+}
+
+void lcd_set_pixel(int x, int y, int color) {
+  // Row address set  (command 0x2B)
+  write_spi_command(PASET);
+  write_spi_data(x);
+  write_spi_data(x);
+  // Column address set  (command 0x2A)
+  write_spi_command(CASET);
+  write_spi_data(y);
+  write_spi_data(y);
+  // Now illuminate the pixel (2nd pixel will be ignored)
+  write_spi_command(RAMWR);
+  write_spi_data((color >> 4) & 0xFF);
+  write_spi_data(((color & 0xF) << 4) | ((color >> 8) & 0xF));
+  write_spi_data(color & 0xFF);
+}
+
+void lcd_set_line(int x0, int y0, int x1, int y1, int color) {
+  int dy = y1 - y0;
+  int dx = x1 - x0;
+  int stepx, stepy;
+
+  if(dy < 0){
+    dy=-dy;
+    stepy=-1;
+  }else{
+    stepy=1;
+  }
+  if( dx < 0 ){
+    dx=-dx;
+    stepx=-1;
+  }else{
+    stepx=1;
+  }
+  dy <<= 1; // dy is now 2*dy
+  dx <<= 1; // dx is now 2*dx
+
+  lcd_set_pixel(x0, y0, color);
+  if (dx > dy) {
+    int fraction = dy - (dx >> 1);// same as 2*dy - dx
+    while (x0 != x1) {
+      if (fraction >= 0) {
+        y0 += stepy;
+        fraction -= dx; // same as fraction -= 2*dx
+      }
+      x0 += stepx;
+      fraction += dy; // same as fraction -= 2*dy
+      lcd_set_pixel(x0, y0, color);
+    }
+  }else{
+    int fraction = dx - (dy >> 1);
+    while (y0 != y1) {
+      if (fraction >= 0) {
+        x0 += stepx;
+        fraction -= dy;
+      }
+      y0 += stepy;
+      fraction += dx;
+      lcd_set_pixel(x0, y0, color);
+    }
+  }
+
+}
+
+void lcd_set_rect(int x0, int y0, int x1, int y1, unsigned char fill, int color) {
+  int xmin, xmax, ymin, ymax;
+  int i;
+  // check if the rectangle is to be filled
+  if (fill == FILL) {
+    // best way to create a filled rectangle is to define a drawing box // and loop two pixels at a time
+    // calculate the min and max for x and y directions
+    xmin = (x0 <= x1) ? x0 : x1;
+    xmax = (x0 > x1) ? x0 : x1;
+    ymin = (y0 <= y1) ? y0 : y1;
+    ymax = (y0 > y1) ? y0 : y1;
+    // specify the controller drawing box according to those limits // Row address set (command 0x2B)
+    write_spi_command(PASET);
+    write_spi_data(xmin);
+    write_spi_data(xmax);
+    // Column address set  (command 0x2A)
+    write_spi_command(CASET);
+    write_spi_data(ymin);
+    write_spi_data(ymax);
+    // WRITE MEMORY
+    write_spi_command(RAMWR);
+    // loop on total number of pixels / 2
+    for (i = 0; i < ((((xmax - xmin + 1) * (ymax - ymin + 1)) / 2) + 130); i++) {
+      // use the color value to output three data bytes covering two pixels
+      write_spi_data((color >> 4) & 0xFF);
+      write_spi_data(((color & 0xF) << 4) | ((color >> 8) & 0xF));
+      write_spi_data(color & 0xFF);
+    }
+  } else {
+    // best way to draw un unfilled rectangle is to draw four lines
+    lcd_set_line(x0, y0, x1, y0, color);
+    lcd_set_line(x0, y1, x1, y1, color);
+    lcd_set_line(x0, y0, x0, y1, color);
+    lcd_set_line(x1, y0, x1, y1, color);
+  }
+}
+
+void lcd_set_circle(int x0, int y0, int radius, int color) {
+  int f = 1 - radius;
+  int ddF_x = 0;
+  int ddF_y = -2 * radius;
+  int x = 0;
+  int y = radius;
+  lcd_set_pixel(x0, y0 + radius, color);
+  lcd_set_pixel(x0, y0 - radius, color);
+  lcd_set_pixel(x0 + radius, y0, color);
+  lcd_set_pixel(x0 - radius, y0, color);
+
+  while(x < y) {
+    if(f >= 0) {
+      y--;
+      ddF_y += 2;
+      f += ddF_y;
+    }
+    x++;
+    ddF_x += 2;
+    f += ddF_x + 1;
+    lcd_set_pixel(x0 + x, y0 + y, color);
+    lcd_set_pixel(x0 - x, y0 + y, color);
+    lcd_set_pixel(x0 + x, y0 - y, color);
+    lcd_set_pixel(x0 - x, y0 - y, color);
+    lcd_set_pixel(x0 + y, y0 + x, color);
+    lcd_set_pixel(x0 - y, y0 + x, color);
+    lcd_set_pixel(x0 + y, y0 - x, color);
+    lcd_set_pixel(x0 - y, y0 - x, color);
+  }
+}
+
+void lcd_put_char(char c, int x, int y, int size, int fcolor, int bcolor) {
+  /*extern const unsigned char FONT6x8[97][8];
+  extern const unsigned char FONT8x8[97][8];
+  extern const unsigned char FONT8x16[97][16];*/
+  int i,j;
+  unsigned int nCols;
+  unsigned int nRows;
+  unsigned int nBytes;
+  unsigned char PixelRow;
+  unsigned char Mask;
+
+  unsigned int Word0;
+  unsigned int Word1;
+  unsigned char *pFont;
+  unsigned char *pChar;
+  unsigned char *FontTable[] = {(unsigned char *)FONT6x8, (unsigned char *)FONT8x8, (unsigned char *)FONT8x16};
+
+  // get pointer to the beginning of the selected font table pFont = (unsigned char *)FontTable[size];
+  // get the nColumns, nRows and nBytes
+  nCols = *pFont;
+  nRows = *(pFont + 1);
+  nBytes = *(pFont + 2);
+  // get pointer to the last byte of the desired character
+  pChar = pFont + (nBytes * (c - 0x1F)) + nBytes - 1;
+  // Row address set  (command 0x2B)
+  write_spi_command(PASET);
+  write_spi_data(x);
+  write_spi_data(x + nRows - 1);
+  // Column address set  (command 0x2A)
+  write_spi_command(CASET);
+  write_spi_data(y);
+  write_spi_data(y + nCols - 1);
+
+  // Write Memory
+  write_spi_command(RAMWR);
+
+  // loop on each row, working backwards from the bottom to the top for (i = nRows - 1; i >= 0; i--) {
+  Mask = 0x80;
+  for (i = nRows - 1; i >= 0; i--) {
+    // copy pixel row from font table and then decrement row
+    PixelRow = *pChar--;
+    // loop on each pixel in the row (left to right)
+    // Note: we do two pixels each loop
+    Mask = 0x80;
+    for (j = 0; j < nCols; j += 2) {
+      // if pixel bit set, use foreground color; else use the background color
+      // now get the pixel color for two successive pixels
+      if((PixelRow & Mask) == 0) {
+        Word0 = bColor;
+      } else {
+        Word0 = fColor;
+      }
+      Mask = Mask >> 1;
+      if ((PixelRow & Mask) == 0) {
+        Word1 = bColor;
+      } else {
+        Word1 = fColor;
+      }
+      Mask = Mask >> 1;
+      // use this information to output three data bytes
+      write_spi_data((Word0 >> 4) & 0xFF);
+      write_spi_data(((Word0 & 0xF) << 4) | ((Word1 >> 8) & 0xF));
+      write_spi_data(Word1 & 0xFF);
+    }
+  }
+}
+
+void lcd_put_str(char* string, int x, int y, int size, int fcolor, int bcolor) {
+  // loop until null-terminator is seen
+  while (*pString != 0x00) {
+    // draw the character
+    lcd_put_char(*pString++, x, y, size, fColor, bColor);
+    // advance the y position
+    if (size == SMALL) {
+      y = y + 6;
+    } else if (size == MEDIUM) {
+      y = y + 8;
+    } else {
+      y = y + 8;
+    }
+    // bail out if y exceeds 131
+    if (y > 131) break;
+  }
+}
+
 
 /* USER CODE END PFP */
 
@@ -64,6 +649,9 @@ int main(void)
 {
 
   /* USER CODE BEGIN 1 */
+
+  init_spi();
+  init_lcd();
 
   /* USER CODE END 1 */
 
@@ -81,6 +669,11 @@ int main(void)
   MX_USART4_UART_Init();
 
   /* USER CODE BEGIN 2 */
+
+  lcd_clear_screen();
+  HAL_Delay(10);
+  lcd_put_str("This shit is working", 65, 65, MEDIUM, BLACK, WHITE);
+  HAL_Delay(10);
 
   /* USER CODE END 2 */
 
@@ -148,10 +741,10 @@ void Error_Handler(void)
 {
   /* USER CODE BEGIN Error_Handler */
   /* User can add his own implementation to report the HAL error return state */
-  while(1) 
+  while(1)
   {
   }
-  /* USER CODE END Error_Handler */ 
+  /* USER CODE END Error_Handler */
 }
 
 #ifdef USE_FULL_ASSERT
@@ -176,10 +769,10 @@ void assert_failed(uint8_t* file, uint32_t line)
 
 /**
   * @}
-  */ 
+  */
 
 /**
   * @}
-*/ 
+*/
 
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
